Initial model to test the hypothesis that bilingualism data can predict the cognitive Flanker Effect.
A simple Linear Regression model was trained and evaluated. Feature selection was performed to improve performance:

- The initial model with 5 features performed poorly (R² < -0.6).

- Removing features 'AoA' and 'Immqual' (initially five 'Age', 'Otherlang', 'Immersion', 'AoA', 'Immqual', though 'Age' doesn't have a significant effect on model score) reduced noise, improved the score, and simplified the model.

- Polynomial regression (degree=2) was used but resulted in a poorer RMSE score than previous.

The final model yields an R² of -0.14 and an RMSE of 14.15 ms. These results suggest that bilingualism alone is not a strong linear predictor of the Flanker Effect.

Redoing this because of the previous issue with the .ipynb file being in .venv